Move re-implementation to own files
