+++ phial/__init__.py
@@ -1,8 +1,6 @@
 from phial.bot import Phial
-from phial.globals import command, g
+from phial.globals import command
 from phial.scheduler import Schedule
-from phial.wrappers import (Response, Attachment, MessageAttachment,
-                            MessageAttachmentField)
+from phial.wrappers import Response, Attachment
 
-__all__ = ['Phial', 'command', 'Response', 'Attachment', 'MessageAttachment',
-           'MessageAttachmentField', 'g', 'Schedule']
+__all__ = ['Phial', 'command', 'Response', 'Attachment', 'Schedule']
