+++ phial/globals.py
@@ -1,24 +1,15 @@
 from werkzeug.local import LocalStack, LocalProxy
-from typing import cast, Dict
-from phial.wrappers import Command
+from typing import cast
+from phial.wrappers import Message
 
 
-def _find_command() -> Command:
+def _find_command() -> Message:
     '''Gets the command from the context stack'''
     top = _command_ctx_stack.top
     if top is None:
         raise RuntimeError('Not in a context with a command')
-    return cast(Command, top)
-
-
-def _get_global() -> Dict:
-    top = _global_ctx_stack.top
-    if top is None:
-        raise RuntimeError('Working outside the app context')
-    return cast(Dict, top)
+    return cast(Message, top)
 
 
 _command_ctx_stack = LocalStack()  # type: ignore
-_global_ctx_stack = LocalStack()  # type: ignore
-command = cast(Command, LocalProxy(_find_command))  # type: Command
-g = cast(Dict, LocalProxy(_get_global))  # type: Dict
+command: Message = cast(Message, LocalProxy(_find_command))
