+++ phial/types.py
@@ -1,0 +1,6 @@
+from typing import TYPE_CHECKING, Union
+
+if TYPE_CHECKING:
+    from phial.wrappers import Response, Attachment  # noqa
+
+PhialResponse = Union[None, str, 'Response', 'Attachment']
