+++ phial/utils.py
@@ -1,4 +1,6 @@
 import re
+from phial.wrappers import Message
+from typing import List, Dict, Optional
 
 
 def parse_help_text(help_text: str) -> str:
@@ -16,3 +18,20 @@ def parse_help_text(help_text: str) -> str:
     help_text = re.sub(r'(^[ \t]+|[ \t]+)', ' ', help_text, flags=re.M)
 
     return help_text
+
+
+def parse_slack_output(slack_rtm_output: List[Dict]) -> Optional['Message']:
+    output_list = slack_rtm_output
+    if output_list and len(output_list) > 0:
+        for output in output_list:
+            if(output and 'text' in output):
+                bot_id = None
+                if 'bot_id' in output:
+                    bot_id = output['bot_id']
+                return Message(output['text'],
+                               output['channel'],
+                               output['user'],
+                               output['ts'],
+                               output['team'],
+                               bot_id)
+    return None
