+++ phial/wrappers.py
@@ -1,24 +1,46 @@
-from typing import Dict, Pattern, IO, Optional, List, Union
+import re
+from typing import Dict, Pattern, IO, Optional, List, Callable
+from phial.types import PhialResponse
 
-MessageAttachmentJson = Dict[str, Union[str, Union[str, Dict[str, str]]]]
+
+class Response():
+    def __init__(self,
+                 channel: str,
+                 text: Optional[str] = None,
+                 original_ts: Optional[str] = None,
+                 reaction: Optional[str] = None,
+                 ephemeral: bool = False,
+                 user: Optional[str] = None,
+                 attachments: Optional[Dict] = None) -> None:
+        self.channel = channel
+        self.text = text
+        self.original_ts = original_ts
+        self.reaction = reaction
+        self.ephemeral = ephemeral
+        self.user = user
+        self.attachments = attachments
+
+    def __repr__(self) -> str:
+        return "<Response: {0}>".format(self.text)
+
+    def __eq__(self, other: object) -> bool:
+        return self.__dict__ == other.__dict__
+
+
+class Attachment():
+    def __init__(self,
+                 channel: str,
+                 filename: Optional[str] = None,
+                 content: Optional[IO[bytes]] = None) -> None:
+        self.channel = channel
+        self.filename = filename
+        self.content = content
+
+    def __repr__(self) -> str:
+        return "<Attachment in {0} >".format(self.channel)
 
 
 class Message():
-    '''
-    Used by phial to store data about the message that initiated
-    the command.
-
-    Attributes:
-        text(str): The message contents
-        channel(str): The Slack channel ID the message was sent from
-        user(str): The user who sent the message
-        timestamp(str): The message's timestamp
-        team(str): The Team ID of the Slack workspace the message was
-                   sent from
-        bot_id(str, optional): If the message was sent by a bot
-                               the ID of that bot.
-                               Defaults to None.
-    '''
     def __init__(self,
                  text: str,
                  channel: str,
@@ -43,209 +65,50 @@ class Message():
         return self.__dict__ == other.__dict__
 
 
-class Command():
-    '''
-    The command object used by Phial.
-
-    Attributes:
-        command_pattern(str): The regex used for matching the command
-        channel(str): The Slack channel ID the command was called from
-        args(dict): Any arguments passed to the command
-        user(str): The Slack User ID of the user who intiated the command
-        message(`Message`): The message that initiated the command
-        message_ts(str): The timestamp of the message that initiated the
-                         command
-        team(str): The Team ID of the Slack workspace the command was
-                   called from
-    '''
+class Command:
     def __init__(self,
-                 command_pattern: Optional[Pattern[str]],
-                 channel: str,
-                 args: Optional[Dict],
-                 user: str,
-                 message: Message,
-                 team: str) -> None:
-        self.command_pattern = command_pattern
-        self.channel = channel
-        self.args = args
-        self.user = user
-        self.message = message
-        self.message_ts = message.timestamp
-        self.team = team
+                 pattern: str,
+                 func: Callable[..., PhialResponse],
+                 case_sensitive: bool = False,
+                 help_text_override: Optional[str] = None):
+        self.pattern_string = pattern
+        self.pattern = self._build_pattern_regex(pattern, case_sensitive)
+        self.alias_patterns = self._get_alias_patterns(func)
+        self.func = func
+        self.case_sensitive = case_sensitive
+        self.help_text_override = help_text_override
 
-    def __repr__(self) -> str:
-        return "<Command: {0}, {1} in {2}:{3}>".format(self.message.text,
-                                                       self.args,
-                                                       self.channel,
-                                                       self.team)
+    def _get_alias_patterns(self, func: Callable) -> List[Pattern]:
+        patterns: List[Pattern] = []
+        if hasattr(func, 'alias_patterns'):
+            for pattern in func.alias_patterns:  # type: ignore
+                patterns.append(self._build_pattern_regex(pattern))
+        return patterns
 
-    def __eq__(self, other: object) -> bool:
-        return self.__dict__ == other.__dict__
+    @staticmethod
+    def _build_pattern_regex(pattern: str,
+                             case_sensitive: bool = False) -> Pattern[str]:
+        command = re.sub(r'(<\w+>)', r'(\"?\1\"?)', pattern)
+        command_regex = re.sub(r'(<\w+>)', r'(?P\1[^"]*)', command)
+        return re.compile("^{}$".format(command_regex),
+                          0 if case_sensitive else re.IGNORECASE)
 
+    def pattern_matches(self,
+                        message: Message) -> Optional[Dict]:
+        match = self.pattern.match(message.text)
+        if match:
+            return match.groupdict()
 
-class MessageAttachmentField():
-    '''
-    Displays field in message attachment
+        # Only check aliases if main pattern does not match
+        for alias in self.alias_patterns:
+            match = alias.match(message.text)
+            if match:
+                return match.groupdict()
 
-    Attributes:
-        title(str): The title of the field
-        value(str): The value of the field
-        short(bool): If false, 2 fields will be displayed on the same line
+        return None
 
-    '''
-    def __init__(self,
-                 title: Optional[str] = None,
-                 value: Optional[str] = None,
-                 short: Optional[bool] = None) -> None:
-        self.title = title
-        self.value = value
-        self.short = short
-
-    def __repr__(self) -> str:
-        return "<Message Attachment Field with title {0} >".format(self.title)
-
-
-class MessageAttachment():
-    '''
-    Displays a message attachment
-
-    Attributes:
-        fallback(str): Plain text summary of the attachment
-        color(str): Hex code beginning with a '#' for the side bar colour
-        author_name(str): Small text to display the authors name
-        author_link(str): URL for link on authors name
-        author_icon(str): URL for 16x16 icon beside the authors name
-        title(str): The title of the attachment
-        title_link(str): A URL for the attachment title
-        text(str): The main text body of an attachment
-        image_url(str): URL to an image file that will be displayed with
-                        an attachment
-        thumb_url(str): URL to an image file that will be displayed as a
-                        thumbnail on the right side of a message attachment
-        fields(list): A list of MessageAttachmentFields to display on the
-                      attachment
-        footer(str): Small text to help contextualize and identify an
-                     attachment
-        footer_icon(str): URL for 16x16 icon beside the footer
-
-    '''
-    def __init__(self,
-                 fallback: Optional[str] = None,
-                 color: Optional[str] = None,
-                 author_name: Optional[str] = None,
-                 author_link: Optional[str] = None,
-                 author_icon: Optional[str] = None,
-                 title: Optional[str] = None,
-                 title_link: Optional[str] = None,
-                 text: Optional[str] = None,
-                 image_url: Optional[str] = None,
-                 thumb_url: Optional[str] = None,
-                 fields: Optional[List[MessageAttachmentField]] = None,
-                 footer: Optional[str] = None,
-                 footer_icon: Optional[str] = None) -> None:
-        self.fallback = fallback
-        self.color = color
-        self.author_name = author_name
-        self.author_link = author_link
-        self.author_icon = author_icon
-        self.title = title
-        self.title_link = title_link
-        self.text = text
-        self.image_url = image_url
-        self.thumb_url = thumb_url
-        self.fields = fields
-        self.footer = footer
-        self.footer_icon = footer_icon
-
-    def __repr__(self) -> str:
-        return "<Message Attachment with text {0} >".format(self.text)
-
-
-class Response():
-    '''
-    When returned in a command function will send a message, or reaction to
-    slack depending on contents.
-
-    Attributes:
-        channel(str): The Slack channel ID the response will be sent to
-        text(str): The response contents
-        original_ts(str): The timestamp of the original message. If populated
-                          will put the text response in a thread
-        reation(str): A valid slack emoji name. NOTE: will only work when
-                      original_ts is populated
-        attachments(Union[List[MessageAttachment],
-                          List[Dict[str, Dict[str, str]]]]):
-                          A list of MessageAttachment objects to be attached
-                          to the message
-        ephemeral(bool): Whether to send the message as an ephemeral message
-        user(str): The user id to display the ephemeral message to
-
-    Examples:
-        The following would send a message to a slack channel when executed ::
-
-            @slackbot.command('ping')
-            def ping():
-                return Response(text="Pong", channel='channel_id')
-
-        The following would send a reply to a message in a thread ::
-
-            @slackbot.command('hello')
-            def hello():
-                return Response(text="hi",
-                                channel='channel_id',
-                                original_ts='original_ts')
-
-        The following would send a reaction to a message ::
-
-            @slackbot.command('react')
-            def react():
-                return Response(reaction="x",
-                                channel='channel_id',
-                                original_ts='original_ts')
-
-    '''
-    def __init__(self,
-                 channel: str,
-                 text: Optional[str] = None,
-                 original_ts: Optional[str] = None,
-                 attachments: Optional[Union[List[MessageAttachment],
-                                       List[MessageAttachmentJson]]] = None,
-                 reaction: Optional[str] = None,
-                 ephemeral: bool = False,
-                 user: Optional[str] = None) -> None:
-        self.channel = channel
-        self.text = text
-        self.original_ts = original_ts
-        self.reaction = reaction
-        self.attachments = attachments
-        self.ephemeral = ephemeral
-        self.user = user
-
-    def __repr__(self) -> str:
-        return "<Response: {0}>".format(self.text)
-
-    def __eq__(self, other: object) -> bool:
-        return self.__dict__ == other.__dict__
-
-
-class Attachment():
-    '''
-    When returned in a command function will send an attachment to Slack
-
-    Attributes:
-        channel(str): The Slack channel ID the file will be sent to
-        filename(str): The filename of the file
-        content(`io.BufferedReader`): The file to send to Slack. Open file
-                                      using open('<file>', 'rb')
-
-    '''
-    def __init__(self,
-                 channel: str,
-                 filename: Optional[str] = None,
-                 content: Optional[IO[bytes]] = None) -> None:
-        self.channel = channel
-        self.filename = filename
-        self.content = content
-
-    def __repr__(self) -> str:
-        return "<Attachment in {0} >".format(self.channel)
+    @property
+    def help_text(self) -> Optional[str]:
+        if self.help_text_override is not None:
+            return self.help_text_override
+        return self.func.__doc__
