+++ setup.cfg
@@ -1,2 +1,5 @@
 [metadata]
 license_file = LICENSE
+
+[flake8]
+exclude = .git,__pycache__,docs/source/conf.py,old,build,dist
